Initial Commit
This is the initial launching of my websites Keys & Notes. I aim to
make a website that shows the user information regarding keys, the
notes within the keys, chords and inversions. So far C Major and G
Major is finished.  I have also done a rough draft of the piano I will
be using to visualize the notes and I will add another octave to it
later.